requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
volcenginesdkarkruntime==1.0.0
httpx==0.27.0